Correctly extract the GCP auth parameters
* Extract the correct keys from the GCP `key_json` auth parameter
  provided by the user.
* Restrict the unsubscription from and deletion of topics that were
  not properly created.

+++ components/google-cloud/pubsub.js
@@ -65,10 +65,14 @@ module.exports = {
       const pubSubClient = new PubSub(sdkParams);
 
       const subscriptionName = this.db.get('subscriptionName');
-      await pubSubClient.subscription(subscriptionName).delete();
+      if (subscriptionName) {
+        await pubSubClient.subscription(subscriptionName).delete();
+      }
 
       const topicName = this.db.get('topicName')
-      await pubSubClient.topic(topicName).delete();
+      if (topicName) {
+        await pubSubClient.topic(topicName).delete();
+      }
     },
   },
   async run(event) {
